chore: flatten demo structure
