+++ apps/react-demo/pages/vanilla/[id].tsx
