+++ apps/react-demo/src/products/ui/products-shell.comp.tsx
@@ -3,22 +3,27 @@ import Box from '@material-ui/core/Box';
 import Container from '@material-ui/core/Container';
 import Typography from '@material-ui/core/Typography';
 import { useRouter } from 'next/router';
-import { FunctionComponent } from 'react';
+import React, { FunctionComponent } from 'react';
 import { Link } from 'react-demo/layout/link';
 import { ProductPagination } from 'react-demo/products/models/product-pagination';
+import { Recipe } from 'react-demo/recipes/recipe';
 
 type Props = {
   query: ProductPagination;
+  recipe: Recipe;
 };
 
-export const ProductsShellComp: FunctionComponent<Props> = ({ query, children }) => {
+export const ProductsShellComp: FunctionComponent<Props> = ({ query, recipe, children }) => {
   const router = useRouter();
 
   return (
     <Container maxWidth="lg">
       <Box my={4}>
-        <Typography variant="h4" component="h1" gutterBottom>
-          Products
+        <Typography variant="h4" component="h1">
+          Products - {recipe.title}
+        </Typography>
+        <Typography variant="h6" component="h2" gutterBottom>
+          {recipe.description}
         </Typography>
         {children}
         <div>
